requirements.txt
beautifulsoup4==4.9.0
Pillow==7.0.0
requests==2.23.0
soupsieve==2.0

